blacklisting/requirements.txt
scapy
scapy-http
pyyaml
